Move components to appropriate packages
