+++ client/src/components/event/card-results.js
