+++ client/src/components/event/card-view.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Card from 'react-bootstrap/Card';
 import { Link } from 'react-router-dom';
 import moment from 'moment';
-import Api from '../../../services/api';
+import Api from '../../services/api';
 
 class CardView extends React.Component {
   constructor(props) {
