+++ client/src/components/restaurant/restaurant-item.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Tab, Tabs, Container } from 'react-bootstrap';
 import PropTypes from 'prop-types';
-import Api from '../../../services/api';
+import Api from '../../services/api';
 import About from './restaurant-about';
-import '../../../style/restaurant-item.css';
-import Feedback from '../../feedbacks/feedback';
+import '../../style/restaurant-item.css';
+import Feedback from '../feedbacks/feedback';
 
 
 class Restaurant extends Component {
